Use BufferedSource to modularize stream ops
